+++ src/hevm/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
  - Fix regression in VMTest compliance.
+ - `hevm exec` ergonomics:
+  - Allow code/calldata prefixed with 0x
+  - create transactions with specific caller nonce
 
 ## 0.33 - 2019-08-06
  - Full compliance with the [General State Tests][245] (with the